+++ CHANGELOG.md
@@ -14,6 +14,7 @@ for a function that will run when accepting a new connection.
 ([#210](https://github.com/anmonteiro/lumo/issues/210)).
 - Add `lumo.core/exit` function.
 - Pretty print JavaScript objects & arrays.
+- Add a new `-A / --checked-arrays` command line option that can be `warn` or `error`.
 
 ### Changes
 
@@ -27,6 +28,7 @@ for a function that will run when accepting a new connection.
 - Socket REPL fails to isolate NS ([#158](https://github.com/anmonteiro/lumo/issues/158)).
 - Fix NPM installation as root ([#206](https://github.com/anmonteiro/lumo/issues/206)).
 - Deduplicate the results of `lumo.repl/apropos`.
+- Fix issue that prevented `*print-namespace-maps*` to be `true` on REPL startup.
 
 ## [1.6.0](https://github.com/anmonteiro/lumo/compare/1.5.0...1.6.0) (2017-06-30)